Move T0x1f to own file
