+++ src/Parser.cpp
@@ -1548,36 +1548,6 @@ SymbolBBox Parser::readSymbolBBox()
 }
 
 
-// @todo Probably specifies the 'Package Properties'
-T0x1f Parser::readT0x1f()
-{
-    spdlog::debug(getOpeningMsg(__func__, mDs.getCurrentOffset()));
-
-    T0x1f obj;
-
-    obj.name = mDs.readStringLenZeroTerm();
-
-    std::string unknownStr0 = mDs.readStringLenZeroTerm(); // @todo figure out
-    spdlog::debug("{} unknownStr0 = {}", __func__, unknownStr0);
-
-    obj.refDes = mDs.readStringLenZeroTerm();
-
-    std::string unknownStr1 = mDs.readStringLenZeroTerm(); // @todo figure out
-    spdlog::debug("{} unknownStr1 = {}", __func__, unknownStr1);
-
-    obj.pcbFootprint = mDs.readStringLenZeroTerm();
-
-    // Maybe the last two bytes specify the amount of units the symbols has?
-    // Also called "Section Count"
-    mDs.printUnknownData(2, std::string(__func__) + " - 0 - Prob. Unit Count");
-
-    spdlog::debug(getClosingMsg(__func__, mDs.getCurrentOffset()));
-    spdlog::info(to_string(obj));
-
-    return obj;
-}
-
-
 // @todo create/update Kaitai file
 GeneralProperties Parser::readGeneralProperties()
 {
